// model/shell/parse.model.ts
export interface Pos {
  Line: number;
  Col: number;
  Offset: number;
}
/**
 * Our notion of base node, as opposed to `MvdanSh.BaseNode`.
 */
export interface BaseNode {
  // End: Pos;
  // Pos: Pos;

  /** Single instance for entire parse tree */
  meta: FileMeta;
  /** Reference to parent node  */
  parent: null | ParsedSh;

  /**
   * Sometimes a node needs a uid:
   * - identify command substitution file `/dev/cs-${pid}-${uid}`
   * - TODO prevent > 1 background process per respective subterm 
   */
  uid?: string;

  /** Used for test expansion */
  boolean?: boolean;
  /** Used for arithmetic expansion */
  number?: number;
  /** Used for arithmetic/boolean expansion */
  string?: string;
  /** Used to calculate actual exit codes */
  exitCode?: number;
  /** Used by Assign nodes only */
  declOpts?: Partial<BaseAssignOpts>;
  /** Used by ParamExp nodes only */
  paramDef?: ParameterDef<any, any>;
  /** Used by Redirects only */
  redirDef?: RedirectDef<any>;
  /** Used by ForClause and WhileClause only */
  lastIterated?: number;
}

export type ParsedSh = (
  | ArithmCmd
  | ArithmExp
  | ArrayElem
  | ArithmExpr
  | ArrayExpr
  | Assign
  | BinaryArithm
  | BinaryCmd
  | BinaryTest
  | Block
  | CallExpr
  | CaseClause
  | CaseItem
  | CmdSubst
  | Comment
  | CStyleLoop
  | Command
  | CoprocClause
  | DblQuoted
  | DeclClause
  | ExtGlob
  | File
  | ForClause
  | FuncDecl
  | IfClause
  | LetClause
  | Lit
  | Loop
  | ParamExp
  | ParenArithm
  | ParenTest
  | ProcSubst
  | Redirect
  | SglQuoted
  | Stmt
  | Subshell
  | TestClause
  | TimeClause
  | TestExpr
  | UnaryArithm
  | UnaryTest
  | WhileClause
  | Word
  | WordIter
  | WordPart
);
// ...
export type CallExpr = Sh.CallExprGeneric<BaseNode, Pos, string>
// ...
export interface FileMeta {
  pid: number;
  /** This is a shell iff `pid === sid` */
  sid: number;
  sessionKey: string;
}